# src/swagger_output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Nodejs with TypeORM",
    "description": "My API documentation",
    "contact": {
      "name": "Aykut Harmancı"
    },
    "license": {
      "name": "Aykut Harmancı",
      "url": "https://instagram.com/aykuthrmnc"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/",
      "description": "local server"
    },
    {
      "url": "http://localhost:5000/",
      "description": "the other server"
    }
  ],
  "tags": [
    {
      "name": "User",
      "description": "User endpoint"
    },
    {
      "name": "Auth"
    }
  ],
  "paths": {
    "/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "Kullanıcı girişi.",
        "parameters": [],
        "responses": {
          "400": {
            "description": "Bad Request"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Login"
              },
              "examples": {
                "User1": {
                  "$ref": "#/components/examples/User1"
                },
                "User2": {
                  "$ref": "#/components/examples/User2"
                }
              }
            }
          }
        }
      }
    },
    "/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "Kullanıcı girişi.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "example": "any"
                  },
                  "password": {
                    "example": "any"
                  },
                  "firstName": {
                    "example": "any"
                  },
                  "lastName": {
                    "example": "any"
                  },
                  "phoneNumber": {
                    "example": "any"
                  },
                  "email": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/profile": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "Kullanıcı profili.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "avatar": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/users/": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "Kullanıcıları listeler.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      },
      "post": {
        "tags": [
          "User"
        ],
        "description": "Yeni kullanıcı ekler",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "example": "any"
                  },
                  "password": {
                    "example": "any"
                  },
                  "firstName": {
                    "example": "any"
                  },
                  "lastName": {
                    "example": "any"
                  },
                  "phoneNumber": {
                    "example": "any"
                  },
                  "email": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/users/withRole": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "Kullanıcıları rolleriyle beraber listeler.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "Kullanıcı bilgisini getirir.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID gereklidir."
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      },
      "put": {
        "tags": [
          "User"
        ],
        "description": "Kullanıcı bilgilerini günceller.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID gereklidir."
          }
        ],
        "responses": {
          "200": {
            "description": "User registered successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateUser"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateUser"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "example": "any"
                  },
                  "password": {
                    "example": "any"
                  },
                  "firstName": {
                    "example": "any"
                  },
                  "lastName": {
                    "example": "any"
                  },
                  "email": {
                    "example": "any"
                  },
                  "phoneNumber": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "User"
        ],
        "description": "Kullanıcıyı siler.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID gereklidir."
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    }
  },
  "components": {
    "examples": {
      "User1": {
        "value": {
          "username": "aykuthrmnc",
          "password": "123456"
        },
        "summary": "Ana hesap"
      },
      "User2": {
        "value": {
          "username": "aykuthrmncc",
          "password": "123456"
        },
        "summary": "Yan hesap"
      }
    },
    "schemas": {
      "Login": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": ""
          },
          "password": {
            "type": "string",
            "example": ""
          }
        },
        "xml": {
          "name": "Login"
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "example": ""
          },
          "lastName": {
            "type": "string",
            "example": ""
          },
          "username": {
            "type": "string",
            "example": ""
          },
          "password": {
            "type": "string",
            "example": ""
          },
          "email": {
            "type": "string",
            "example": ""
          },
          "phoneNumber": {
            "type": "string",
            "example": ""
          }
        },
        "xml": {
          "name": "User"
        }
      },
      "CreateUser": {
        "xml": {
          "name": "CreateUser"
        },
        "$ref": "#/components/schemas/User"
      },
      "UpdateUser": {
        "xml": {
          "name": "UpdateUser"
        },
        "$ref": "#/components/schemas/User"
      }
    },
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization"
      }
    }
  },
  "security": {
    "api_key": []
  }
}